+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## UNRELEASED (0.14.2)
 
+✨ **New features**
+
+- New method: `BigRational::clamp()`
+
 👌 **Improvements**
 
 - `BigInteger::modInverse()` now accepts `BigNumber|int|float|string` instead of just `BigInteger`